Remove timeout from airflow.utils and update core usages accordingly
